rename(domain): AbsenceLevel -> AbsencePenalty 이름 변경
