+++ src/main/java/domain/AbsencePenalty.java
@@ -1,6 +1,6 @@
 package domain;
 
-public enum AbsenceLevel {
+public enum AbsencePenalty {
     OUT("제적", 5),
     MEETING("면담", 3),
     WARNING("경고", 2),
@@ -10,12 +10,12 @@ public enum AbsenceLevel {
     private final String level;
     private final int standard;
 
-    AbsenceLevel(String level, int standard) {
+    AbsencePenalty(String level, int standard) {
         this.level = level;
         this.standard = standard;
     }
 
-    public static AbsenceLevel findAbsenceLevel(int absentCount, int lateCount) {
+    public static AbsencePenalty findAbsenceLevel(int absentCount, int lateCount) {
         absentCount += (lateCount / LATE_TO_ABSENCE_THRESHOLD);
         if (absentCount > OUT.standard) {
             return OUT;
